refactor browserCapabilities
